Update Rippl to Swift 5, add SPM support, remove empty tests
