+++ Rippl-Demo/AppDelegate.swift
