+++ Rippl-Demo/Assets.xcassets/Contents.json
