+++ Rippl-Demo/Info.plist
