+++ Rippl-Demo/ViewController.swift
