+++ Sources/Rippl-iOS.h
